instagram_publisher/requirements.txt
httpx>=0.25.0
aiofiles>=23.2.1
pillow>=10.0.0
motor>=3.3.1
python-dotenv>=1.0.0
asyncio>=3.4.3 